Removed implicit conversion from ClassTypeReference to string.
Implicit conversion should be used when object is somehow compatible with the destination type. ClassTypeReference definitely does not feel like a string. Moreover, _classRef can be accessed using GetClassRef anyway.

MSDN recommendation on implicit:
Use it to enable implicit conversions between a user-defined type and another type, if the conversion is guaranteed not to cause a loss of data.